requirements.txt
pytest==7.3.1
selenium==4.9.0
allure-pytest==2.13.2
pytest-xdist==3.2.1
webdriver_manager==3.8.6
termcolor==2.3.0
pytest-selenium==4.0.0

requests==2.29.0